src/pyproject.toml
[tool.poetry]
name = "bot-constructor"
version = "0.1.0"
description = ""
authors = ["Artem Vol <artemvol1998@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
python-telegram-bot = "^21.10"
pydantic = "^2.10.6"
pydantic-settings = "^2.8.0"
orjson = "^3.10.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
